+++ src/pytest_djangoapp/fixtures/__init__.py
@@ -7,6 +7,7 @@ from .messages import messages
 from .migrations import check_migrations
 from .request import request_client, request_factory, request_get, request_post
 from .settings import settings
+from .signals import signals
 from .templates import template_context, template_render_tag, template_strip_tags
 from .users import user, user_create, user_model
 from .utils import conf_app_name, registered_urls, run_app